backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
pandas
openpyxl
python-dotenv
openai
